+++ client/src/components/constants/postTypes.ts
@@ -1,0 +1,24 @@
+export const postTypes = [
+  'All',
+  'Events',
+  'Clubs',
+  'Textbooks',
+  'Roommates',
+  'Opportunities',
+];
+
+export const typeLabels: { [key: string]: string } = {
+  Events: 'Capacity',
+  Clubs: 'Capacity',
+  Textbooks: 'Available',
+  Roommates: 'Looking for',
+  Opportunities: 'Positions Available',
+};
+
+export const typeButtonLabels: { [key: string]: string } = {
+  Events: 'Check In',
+  Clubs: 'Apply',
+  Textbooks: 'Interested',
+  Roommates: 'Interested',
+  Opportunities: 'Apply',
+};
